chore: move star rating to ui directory
